Add task to override Openresolv config
To set the local DNS instance as the primary resolver we need to
override one line in the resolvconf.conf file that is read by
Openresolv, which in turn is the program that updates the resolv.conf
file.

One problem with this approach is that if 127.0.0.1 is present as any
of the nameservers provided in the config, then only that one will be
written to the resolv.conf file. Why that is I don't know.
This will only be an annoyance if Bind fails to start, as then we need
to manually modify this file to get DNS resolution to work again until
Bind is functional.